+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 1.0.1 (in development)
+## 1.0.1 (Sep 8, 2024)
 
 - Added tenant scope to `Entry::findSiblings()`
 - Fixed `tenant_id` validation for strings